Finished day3.
Note that i shot myself a bit in the foot with the solution for part1.
Since I discarded all the information that is necessary for part2
I had to calculate the distance for every point from the beginning again.
Luckily the problem is small enough so that it doesn't matter much.